rename: CardNickName 폴더 생성
